[refactor]move controllers to folder
